requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-dateutil
pytz
email-validator
python-multipart